fixied command argument

+++ src/index.js
@@ -37,6 +37,12 @@ function isvalidkey(key) {
     }
     return entry;
 }
+function check(parts, required){
+    return parts.length===required;
+}
+function isvalid(value){
+    return !isNaN(value)&& Number(value)>=0; 
+}
 setInterval(() => {
      let changed=false;
     for (const [key, entry] of database) {
@@ -62,13 +68,22 @@ const server = net.createServer((c) => {
         const parts = command.split(' ');
 
         if (parts[0] === 'SET') {
+             if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const value = parts[2];
+            
             database.set(key, { type: "string", value: value, expiry: null });
             saveDatabase();
             c.write('OK\r\n');
         }
         else if (parts[0] === 'GET') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
 
@@ -83,6 +98,10 @@ const server = net.createServer((c) => {
 
         }
         else if (parts[0] === 'DEL') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const value = database.delete(key);
 
@@ -95,6 +114,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'EXISTS') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
 
             const entry = isvalidkey(key)
@@ -109,7 +132,10 @@ const server = net.createServer((c) => {
 
         }
         else if (parts[0] === 'KEYS') {
-
+    if (!check(parts, 1)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const value = database.keys()
 
 
@@ -128,8 +154,6 @@ const server = net.createServer((c) => {
                 c.write(0 + ' \r\n');
             }
         }
-
-
         else if (parts[0] === 'FLUSHALL') {
             database.clear();
             saveDatabase();
@@ -137,8 +161,16 @@ const server = net.createServer((c) => {
         }
 
         else if (parts[0] === 'EXPIRE') {
+            if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const time = parts[2];
+            if(!isvalid(time)){
+                c.write('Invalid expire time');
+                return;
+            }
             const currentTime = Date.now();
             const entry = database.get(key);
             if (entry) {
@@ -164,6 +196,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'TTL') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
 
@@ -179,6 +215,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'PERSIST') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = database.get(key);
             const currentTime = Date.now();
@@ -206,6 +246,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'LPUSH') {
+            if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const value = parts[2];
             const entry = database.get(key);
@@ -227,10 +271,21 @@ const server = net.createServer((c) => {
         }
 
         else if (parts[0] === 'LRANGE') {
+            if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const start = Number(parts[2]);
             const end = Number(parts[3]);
-
+           if(!isvalid(start)){
+              c.write('invalid range\r\n');
+                   return;
+           }
+           if(isNaN(end) ||end<-1){
+             c.write('invalid range\r\n');
+            return;
+           }
             const entry = isvalidkey(key);
 
             if (entry) {
@@ -255,6 +310,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'LPOP') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
             if (entry) {
@@ -278,6 +337,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'RPOP') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
             if (entry) {
@@ -301,6 +364,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'RPUSH') {
+            if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const value = parts[2];
             const entry = isvalidkey(key);
@@ -323,6 +390,10 @@ const server = net.createServer((c) => {
 
         }
         else if (parts[0] === 'LLEN') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
 
@@ -337,6 +408,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'HSET') {
+            if (!check(parts, 4)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const field = parts[2];
             const data = parts[3];
@@ -360,6 +435,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] == 'HGET') {
+            if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const field = parts[2];
             const entry = isvalidkey(key);
@@ -386,6 +465,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'HDEL') {
+            if (!check(parts, 3)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const field = parts[2];
             const entry = database.get(key);
@@ -413,6 +496,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'HGETALL') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
             if (entry) {
@@ -431,6 +518,10 @@ const server = net.createServer((c) => {
             }
         }
         else if (parts[0] === 'HLEN') {
+            if (!check(parts, 2)) {
+              c.write('Err wrong number of arguments\r\n');
+             return;
+           }
             const key = parts[1];
             const entry = isvalidkey(key);
